Add issue and pull request templates
GitHub form templates provide structured forms for contributors to
submit bug reports and feature requests. The pull request template
provides guidance on the expected format for pull request titles and
descriptions.

Also displays a link for contributors to ask questions in GitHub
discussions if it is more suitable for their use than a bug report or
feature request.